This should fix input output regardless of magic quotes settings.
NOTE" due to SMTP/Postfix mail provider problems I haven't been able to test it with mail. Please do so and let me know.

You'll see a long list of commented code, showing how PHP is not to good at dealing with higher order functions and the likes. That's why it took so long. Mail for the core list coming.

git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-1.5@912 8555b757-d854-4b86-925a-82fd84c90ff4