Add watchdog to look for stalled/broken upload clients.
This allows to handle network hiccups and such in a better way (before upload became never-ending).

A client is considered to be broken if :

  - it has spent more than 10 times the average of the last 5 chunk upload time without signaling completion (pausing upload handled)
  - or it is more than 7 chunks late behind the global clients progress average (obviously only makes sense in terasender mode ...)

In such a case the upload is automatically restarted from the last safe point, with a maximum of 3 automatic restarts.

If more than 3 automatic restarts occured the user is notified and asked what to do, he can :

  - retry anyway (reset counters)
  - stop upload alltogether
  - ignore because detection is a false-positive

This process is a bit complex but in a normal situation automatic restart does not even happend so it should be fine.



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3353 8555b757-d854-4b86-925a-82fd84c90ff4