Refactor translation ids, improve naming.
As of now most if not all translations currently used in the code are done for english language.

Language file has also been categorized for better readability.

A total of 59 translations out of 324 have been identified asautomatically  transferable from 1.5 version, this will leave 265 strings to be translated (in addition of emails) ...



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3375 8555b757-d854-4b86-925a-82fd84c90ff4