Add a configurable user profile page (disabled by default).
The goal is to create a framework for user preferences managment.

To enable it set the "user_page" configuration parameter to an array containing <preference_id> : <view_mode> pairs.

View modes can be "read", "write" or false (same as not set, that is will not be displayed).

Supported preferences are :

  - lang (requires lang_userpref_enabled to true) : supports read and write
  - auth_secret (requires auth_remote_user_enabled), supports only read
  - created, supports only read



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3397 8555b757-d854-4b86-925a-82fd84c90ff4