Adding an option to get a link for download instead of sending to recipient (disabled by default).
When using this an anonymous recipient is created (that is, without email) and its download link is given back to the uploader at the end of upload.

When using this option the user can embbed the link in any mail of his choice (newsletter for example), all downloads are obviously counted as done by the single anonymous recipient.

This mode will be helpful for people wanting to use their FileSender server from another application (like big attachments handling in mail clients).

This option must be enabled in the transfer_options configuration associative array under the name get_a_link (advanced mode is ignored as well as default).

Also adding user id in user page (if enabled).



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3400 8555b757-d854-4b86-925a-82fd84c90ff4