Add transfer lang property which the user may choose at upload time (enabled by transfer_recipients_lang_selector_enabled configuration boolean, false by default).
This lang defaults to the configured default language.

It is used to translate emails sent to the recipients (because you may have a filesender instance in your native language but want to send english emails to your recipients).

Also use the transfer owner language if saved in preferences when sending him notifications.



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3438 8555b757-d854-4b86-925a-82fd84c90ff4