Add automatic reminder feature.
Using transfer_automatic_reminder config parameter it is now possible to
automatically send reminders to recipients that did not download files
when the expiration date is nearing.

transfer_automatic_reminder is a list (or a single) of number of days
before expiry at which a reminder will be sent :

  * $config['transfer_automatic_reminder'] = 2; // A reminder will be sent 2 days before expiry
  * $config['transfer_automatic_reminder'] = array(5, 2); // A reminder will be sent 5 and 2 days before expiry


For this system to be accurate the cronjob should run after 00:00.

Test #1128




git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3767 8555b757-d854-4b86-925a-82fd84c90ff4