Add fine upload tracking for clients that supports it (HTML5).
This allows for a more-than-once-per-chunk progressbar update,
this is especially usefull for slow connections as the progressbar
looked "stuck" between 2 chunks before, now the progressbar is updated
as often as the browser reports an upload progress  (varies, seems to
be at least every 100-200kb or so) and up to 2 times per second (to not
overload the browser with gui updates).

This is automatically available (no config).



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3787 8555b757-d854-4b86-925a-82fd84c90ff4