[security fix] Remove remote user config from data returned by user endpoint.
This fix adds an user consent step in the process of obtaining user's remote
access config.

It was possible to get an user's remote access config (if enabled and secret set)
without the user noticing by using a carefully crafted page including a jsonp
request to FileSender's web service and inviting the user to visit it with hope
that he had an open FileSender session.

Now the remote application that want to get access must invite the user to
go to <filesender_url>?s=user&remote_auth_sync_request=<remote application name>
where he will find a code to give back to the requesting application.

The requesting application will then be able to retreive the user's remote
config using the code.

The code is valid for 2 minutes after being generated, only one code can be
valid at the same time.



git-svn-id: https://subversion.assembla.com/svn/file_sender/filesender/branches/filesender-2.0@3797 8555b757-d854-4b86-925a-82fd84c90ff4