Fix test cases.
Fix all salvageable testcases so that they work on a default FileSender install. Should work on Travis as well.

The ReporterTest was unsalvageable, because it uses non-existing API.
The actual reporter does mailing and reporting in one step and is therefore not testable.

The MailTest sends mail to one of the developers, so it shouldn't be enabled automatically.

The DBObjectTest only succeeds if there are prior transfers in the database, which means that the test is not isolated.

These tests have been disabled in config_tests_filesender.xml.