Fix the random number generator and generate valid UUIDs
In order to understand this commit, you need a bit of history.

During the PINE security audit in 2015, a vulnurability was found in the way FileSender generates random identifiers (2.6.13);
PINE found that `unid_id` and `mt_rand` by themselves weren't sufficient and recommended using HMAC for random number generation.
The new implementation thus used `mt_rand` to generate a random HMAC key the first time this was needed,
and all subsequent random numbers would consist of a random number from `mt_rand`, HMAC-ed with this stored random value.
This implementation was never security checked as far as I know.

In 2017, a new security audit was conducted by Radically Open Security.  They found the same code using `mt_rand`, and acknowledged the HMAC in the code (FIS-001), but recommended that these be replaced by newer PHP functions that weren't available during the previous audit.

This was then implemented (PR#54) by adding a polyfill for the newer PHP functions and replacing occurrences of `mt_rand` with `random_int`, which is not in the spirit of the recommendation in FIS-001.

This pull request seeks to remedy FIS-001 the way that was proposed, by replacing the logic for generating random identifiers, instead of simply replacing a function call.