Remove redundant filename check (#160)
As part of the effort to remedy FIS-007 (PR#64), a filename check was introduced.
However, another filename check was already in place.  The new filename check is better,
because it uses the configuration setting to decide which filename is good and which isn't.

This commit removes the old filename check.