Cleanup, remove unused code
There are some references to an instance ID in the code, but these are not needed for FileSender operation.
My best guess is that this code was used for preventing cache, which is useful only during development.
It's possible that this ID was needed for some future feature, but if that's the case that feature is now long forgotten.

By removing all instance ID code, the generateRandomHexString() function is no longer needed and can be removed as well.