storage_filesystem_hashing fixes: use path, nfs volumes, missing directories
This fixes a number of issues that seem to have crept into the
storage_filesystem_hashing use. Specifically, files were not stored
into the created subdirectories, NFS support was lacking (df output
would throw an exception), and it seems to have been expected that the
entire subdirectory tree existed in one code path when it is created
on an on demand basis.

There are likely more to come as I retest this functionality but this
is a good first handful of updates.