Config validation: fail very fast for bad configuration settings in guest
As has been reported in issue 197, if you have a configuration for
guest_options that uses an invalid key then you will have only
discovered this when a guest voucher sent out is attempted.

This PR moves the detection of an invalid option to a top level event.
You will not see a filesender website if there is an invalid setting
but will get a log message telling you about this, what key is
offending, and what the possible keys are in case you have mistyped
something.