Race condition of file.complete and transaction.complete
The patch is mostly comment, better to be informative there than here.

The basic version is that there was a race between calling
file.complete and transaction.complete. Logging can show that for the
final two chunks nothing was explicitly stopping transaction.complete
from landing before the last chunk of the last file. A test was
already stopping transaction.complete from being called more than
once. There are multiple timeout delays involved in eventually calling
transaction.complete so the chances of that happening on a LAN were
less. Normally on a LAN the workers would themselves complete and call
file.complete before the timeouts ran and the already scheduled
transaction.complete would actually call the server.

This patch makes it explicit that transaction.complete is only
scheduled for the final chunk uploaded on the final file.

So I guess like many races, the finding and reasoning largely
outweighs the size of the code patch required to mitigate the
scenario.