stall: new mode to show more info to user (#211)
* stall: new mode to show more info to user

A new mode is available which shows the duration that a chunk has been
active in the upload page. This way a user can see a chunk that is
taking a long time and also get more information that a transfer is
not stalled even though it is a huge file and the progress meter will
move very slowly.

It is early days for the feature. Hopefully the insights gained will
help work out the final kinks in the stall detection and also offer
the user the ability to report a stall that they think is invalid.

* docs