crypto and resume: option to not remember transfer if if crypto is on (#218)
This only effects when you break an upload and then enter the Upload
page again. Normally a dialog "Restart failed transfer" appears
offering to setup the transfer again and "Restart" instead of start
the transfer.

This could be problematic if the user enters a different password for
the resume, having part of the upload encoded with one password and
part with another. It was also reported on
https://github.com/filesender/filesender/issues/80 that resumed
crypted uploads were having issues, so until that is tested to a
greater extent it is better to offer the abilitiy to not allow resume
for crypted transfers.