stall detection: monitoring bytes within chunk transfers
The existing stall detection monitors whole chunks. This code
compliments the existing stall detection by monitoring how many bytes
have been transferred and how long it was since the last "subchunk" of
bytes has been transferred over.

By looking at the xhr progress messages like this we can much more
quickly detect when the single (non ts) or ts upload is starting to
struggle. For example, if you force a stall of say, 10 seconds, on the
server and set the following the worker displays will turn red after 5
seconds elapsed with no new bytes sent over.

$config['upload_considered_too_slow_if_no_progress_for_seconds'] = 5;
$config['upload_display_per_file_stats'] = true;

It is useful to know the low level health of each subchunk so that
network issues that might cause the server to timeout on an xhr but
not be reported by the operating system as an xhr failure are detected
quickly and action such as informing the user and restarting lagging chunks
might be considered in a future PR.