Limits: ability to set maximum file size for encrypted files
This is checked client and server side. One little twist on the client
side is that enabling and disabling encryption will have a toggle
effect on if files are valid.

The code sort of assumes that if the client side transaction.addFile()
doesn't work then it will not magically start to work. The max file
size stuff toggles so that you can enable encryption and see an error
for file too large, but if you then disable encryption again you
should see that the file is ok again.