Shredding: new clearing house to shred deleted files
Running shred(1) directly in the
storage_filesystem_file_deletion_command can make the Web interface
appear to be broken at times. This is due to the execution of shred
being done synchronously.

This new code creates a clearing house for shredding. When a file is
deleted through the Web interface it is moved and marked for shredding.
See the new docs shredding page at docs/v2.0/shredding/index.md
for details.

There are many design options for this. Having a database table allows
failures to be tracked which might not otherwise be noticed. The
original file names are scrubbed so the files to be shredded does not
expose the user who used to own the files.