automatic resume: if a transfer has failed try to resume a few times
When an error would previously lead to a dialog box appearing with a
failed transfer we now try to resume the upload after a period of time
a certain number of times.

This is to help issues like machine going to sleep or wifi dropping
out for a small period of time.

The goal is to maybe see that a resume happened a few times but that
the overall transfer was still completed.