encryption: new Key derivation using importKey then deiveKey
This is an update for how an encryption key is created from the given
password when encrypting or decrypting files. As the manner in which
the key is created from the password can create different keys from
the same password this PR also includes versioning for encrypted
files. This way, files encrypted with the old key derivation will
continue to be decrypted while all new files will switch to the new
key derivation scheme.

This PR moves the code closer in style to the example of the
SubtleCrypto API that is published by mozilla.org at
https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/deriveKey#Example

Key type 1 forces the encryption to AES-CBC at 256 bit and the number
of hashing rounds is chosen to be secure through around 2020. This is
a trade off where I think the magic numbers will have to be brought
out because there is simply no "one size fits all" for the hash
iteration count. If a site is dealing with sensitive material it may
need 10 times the iterations to meet the site policy. I have left this
setting inline for now.

This is to address the original report of the issue in
https://github.com/filesender/filesender/issues/338

Note that the salt is generated server side and sent out to the client
for both initial encryption and made available for decryption. This
does not allow the server any real help in decrypting the data but
does make each transfer of the exact same data with the exact same
password give different results when encrypted. As such it should help
to stifle password hash tables and the like as each individual
transfer will need to be attacked.

The provision of salt moving server to client in the new
encryption_details object paves the way for more information to be
sent over such as file IDs or other data for the transaction that
might help bind the password to the transfer and improve security.