GDPR: Allow a user to leave the system
This update offers a button on the My profile page which
allows the user to leave the FileSender instance. Doing this will
delete the user files, invited guests, etc from the database and
the file contents from the filesystem.

This is a bit tricky as the user can still come back using the same
SAML auth in many cases. So the distunction between the data the
FileSender is storing (in database etc) and the login information is
described above the remove account button.