aggregate statistics: totals for some events without user data (#449)
This adds precalculated aggregate statistics to the system. These
allow you to see the number of uploads, downloads, the size of these
and the number of upload and downloads that did not complete.

The statistics are aggregate and do not contain user identifying
information. Because the values are the totals per time interval and
upserted into the database the viewing of statistics does not require
sophisticated calculation and so the new "admin/aggregate statistics"
page should be viewable even on loaded systems. Hopefully.

There is also a provision (disabled by default) for the cron job to
send the aggregate statistics to the developers if you wish to
configure this feature. These sent statistics will also include which
version you are running.