ui: encryption password length check update (#462)
if you are near the length that will make a password valid then you
could type a few characters and have a bouncing "password too short"
message appear. The checking function both sets the message to
hidden/shown with a fade and reads the state of that ui element. The
state is not updated right away so the code may have toggled the state
a few times by reading older data.