Updates for automatic resume of file uploads (#609)
Improved automatic resume for file uploads after a network interruption.
A number of configuration changes are made to set the default timeouts
to be closer to the test environment. Note that upload_considered_too_slow_if_no_progress_for_seconds is for any upload progress, so a 30 second timeout should work
even on a congested network because some bytes should have moved in that time
frame.

The number of seconds since any data was sent is now shown on the
upload page to provide insight to the user. Once that delay for
no data sent reaches
upload_considered_too_slow_if_no_progress_for_seconds then the
upload will have a resume automatically scheduled.

A new reconnect and continue button was added to the upload page to allow
the user to force a reconnection to the server and a resume. This could sort
of be done before by pausing the transfer and clicking on the play button to
resume though the reconnect button is a more heavy operation in that it will
close terasender connections and respawn them in order to "reconnect" which
may be needed if the laptop has changed IP address for example.

I had thought of trying to combine the pause/play and reconnect into a single
interface but they are similar but different enough features that attempts
to bring them into a single user interface would likely lead the user to
trying to second guess what was happening (eg, does this pause/play trigger
a reconnection so that the change in wifi network will be fixed by doing it).
The terasender_failed_after_many_retries message will now present as human
readable when it happens. Failure to start terasender workers such as might
happen if the network happens to be down at the time is now reported in a
much nicer way.

Config changes
from   0 to 30 for upload_considered_too_slow_if_no_progress_for_seconds
from  10 to 50 for automatic_resume_number_of_retries
from 360 to 10 for automatic_resume_delay_to_resume