Fixes to allow running under PHP8 (#1180)
* Change curly braces for square.

This is a drop in replacement - there is no syntactic difference between {} and [] in PHP.
https://wiki.php.net/rfc/deprecate_curly_braces_array_access

Curly braces are sometimes used to denote array indicies, but this is a style choice. The runtime does not make any distinction.

* Remove magic_quote_gpc handling

`magic_quotes_gpc` was removed in PHP 5.4: https://www.php.net/manual/en/info.configuration.php#ini.magic-quotes-gpc

`get_magic_quotes_gpc` was replaced with function that always returns false in PHP 5.5 - making this block a contradiction for versions of PHP supported by FileSender (7.3, according to the current documentation at the time of writing).