https://bugs.php.net/bug.php?id=77782 workaround
In some circumstances when using an SAPI that does not run 
in-process with the http server (e.g. fcgid, fpm) PHP fails to 
return some environment variables when they are queried as
getenv($foo) - however they can be seen with getenv().

According to the bug, the root cause is that different code
paths are taken when calling getenv() vs getenv($foo):

"... it looks like calling `getenv` without a param eventually
calls fcgi_loadenv with cgi_php_load_env_var, but if you pass 
varname to getenv it only calls fcgi_getenv."

This results in FileSender being unable to load attributes set
by Shibboleth, despite them being present in memory.

Querying the whole array, and then slicing out the entry needed
should be a drop in replacement with equivalent functionality
for all SAPI's.