Make default date format, and that of en & nl, less ambiguous (#1414)
* Make default date format, and that of en & nl, less ambiguous

The default date format of Filesender is Australian. This means
that a date will be written as 05/04/2023. Receiving an English
language email with that date is very ambiguous to me since there's
different interpretations between English speaking countries about
which part is the month in that date. Most notably the current
format is different from what is used in the US which some might
assume, justifiably or not, to be the default for things.

To disambiguate it once and for all, change the default to
05 Apr 2023, which is instantly clear and requires no interpretation.
The exact month abbreviation is based on the locale which is set
right before interpolation.

I would leave it up to translators of the other languages to
make any changed they think are appropriate.

* fix test