Handle when get a link is not available. (#1669)
In that case we might as well select the email option as it is the
only choice. This is an update to
https://github.com/filesender/filesender/issues/1621
because in the past you could proceed without selecting anything and
would "get a link" by default. Now I am automatically selecting the
only option if get a link is not available.

Note that the class get_a_link_top_selector is there for now because
get a link is reshown in the options and thus the HTML ID get_a_link
might not be unique. When get a link is not shown in the main options
then this could be updated to directly use the ID instead.