get a link is a top element only and not in options anymore (#1670)
IIRC there was some magic in the old dev3 to not reshow get a link in
the options though that might have bit rotted over the years.

I have removed it from the options that are shown and made the rest
of the code check if getting a link is an option and if so if it is
selected at the top of the page. Basically making that specific toggle
more of a global than just another option and only presented in one
place in the UI.

This may break other things but it is probably best to do that now and
as other ripples are found I can address them.