Allow generated passwords to not have PBKDF2 performed on them (#1672)
It is redundant to perform PBKDF2 on a randomly generated password
of sufficient length. Given that this was a feature introduced in 2019
on the advice of a professional cryptographer I think it should be
retained until explicitly removed.

Allowing for generated passwords to avoid PBKDF2 allows for secure
site setups. For example, large numbers of PBKDF2 rounds can be set
which will require significant time to perform and obtain the
commensurate security under attack. As generated passwords do not
require PBKDF2 the user has the choice of a custom password and a
significant delay or a more random generated password and immedidate
upload. For example one might set PBKDF2 rounds such that they require
minutes to perform.

I would also love to retain this for future PGP integration as that
will allow even more secure setups with the generated password
automatically encrypted and sent to the recipient with their
explicitly nominated PGP key.