UI3 guest handling update: email shown, no get a link (#1690)
The full email fields are shown for a guest upload. They are removed
for guests who can only send to the inviting party. That needs
retesting.

For now I have disabled the ability for a guest to perform an upload
and "get a link" to the file. It is a code path that needs
revalidation to allow a guest to upload a file and get a link to it
instead of sending emails etc.