d3uid3 bring back show/hide password option (#1705)
We can always prune it off again. If we are doing that for design we
should be aware and accept the security degradation in any non single
person office that is involved in an upload. I imagine there are many
shared spaces in higher education where being able to see a password
as it is typed on a keyboard is much harder than just taking a photo
of it on the screen.