d3ui3 Guest table filtering is in database Guest::fromUserAvailable (#1711)
There is a Guest::fromUserAvailable method to allow the database to
filter guests. If the guest table page does not impose a policy
on the guests passed to it then the surrounding page can use
or extend the database filtering/query code to find what it is wishing
to display.