d3ui3 Bring back pause, play, stop on upload as option
For countries, cities, or towns with slower internet access it is
extremely useful to be able to pause an upload so that a long upload
does not block voip or video calls.

Even if an NREN has significant bandwidth expectations for most users
some of those users may travel to areas with limited bandwidth and may
then appreciate the ability to pause an upload.

At any rate, as maintainer I find it useful enough for myself and
users to have made it at least available as an option. I think people
can comprehend the interface and ignore that part unless it is of use.