ui3: SAML if relogin is needed return to details page correctly
Some parameters need to be feed to the ReturnTo parameter in order
for the user to return back to the transfer and guest details page
correctly.

This can be triggered without waiting for a SAML session timeout by
copying the URL, explicit logout. Then paste the original URL again
and FileSender will ask you to authenticate. Without this PR you will
either get an error on the transfer detail page or a guest not found
general error on the guest detail page. With this PR you will see the
expected page as the return from saml will complete properly.