dev3: remove asmcrypto fallback for non Chromuim Edge (#1753)
This is an old fallback which allows filesender key mode 1 to be
used with ie11 and non Chromium backed Edge browsers.

sonarqube flags a number of blocking items in asmcrypto so this
fallback is being removed. Users of Edge should have all been updated
to the Chromuim backed browser by now as this version of Edge was
released in January 2020.