More logging. More docs. Allow 0-9 in CID.
It is likely you might have footer2 for example which it is handy to
be able to use as a cid by extending that regex to include numbers.

Make the getTemplateCIDImagePath method do some logging so the caller
can just ignore bad configurations and not worry about logs.

Expand the docs.