Update known-installs.md - changed description of adding an install
Changed the description of how to add an install to the list, it was not clear that filesender-dev requires you to be subscribed to it before you can post. Changed that item into an invitation to subscribe to the support list filesender-dev and let the wider community know about your install, added a 3rd option to just send the FileSender board an email.