Move development only packages to requre-dev & docs update
The php-cs-fixer is only needed for the rare reforatting of code.
That was seen as something needed at some stage. At any rate that
should only need to be installed on a development machine.

The sauce/sausage package is not used now. The selenium tests are run
through codeception. codeception can handle doing remote code runs.

The sauce/sausage had become untenable as it forced older versions of
phpunit which can not be used in the GHA environment in the modern
day.

The docs are updated as composer will install all items by default but
in an installation env we only want the --no-dev packages installed.