Possible to save storage subpath in database for easier migration. (#2120)
New option to allow the path from storage to be stored per file in the
database. This is the first step towards allowing various settings
like per day buckets and idp in the storage path to be changed and
existing files continue to be downloadable.

This PR only optionally stores the path. A future PR will allow that
information to be used and this allow storage settings to change and
existing files (uploaded with this new
storage_filesystem_explicitly_store_subpath_per_file enabled) to
continue to be downloaded in the face of settings changes.