Ability to modify chunk size and retain existing uploads
With this you should be able to change the config.php
upload_chunk_size (with upload_crypted_chunk_size and
download_chunk_size to allow a valid configuration) and still be able
to download existing uploads with new uploads performed using the
current settings.

This PR has been tested manually and with the new
ConfigurationABCest.php acceptance test. The acceptance test targets
both the normal filesystem and filesystemChunked storage_type and
varies the chunk size in an A,B,A test with uploads in A, downloads in
B then uploads in B and downloads in A for two configurations. These
automatic tests check for unencrypted, encrypted, and encrypted to zip
transfer scenarios. See testConfigurationAB for details.

Some testing has also been performed using a fake S3 backend and chunk
size changes. As in the past I find myself with no credentials to a
real S3 service so can only test with these environments. For this I
have used a zenko/cloudserver server.

The Azure storage was updated inline with the S3 one but has not been
extensively tested for migrations.

The database.php script creates new columns in the Transfer table and
sets their values to the current upload_chunk_size and
upload_crypted_chunk_size values which should allow existing uploads
to work properly after the upgrade.