UI3: Allow these two buttons on the my profile page to be seen
There are many UI elements that have issues. This is one case
where the text was truncated from view by the button. These are not
the sorts of things one would expect to see in the 202x era of web
design.

This is not a perfect soluion to the problem. The new listx
style sort of relies on there being a button on the left and a
description type element on the right. I would have much preferred to
use a toolkit that better supports web forms or not have built things
in a flex just to add some buttons.

This was reported here https://github.com/filesender/filesender/issues/2181