Backport these to allow the database script to find them (#2223)
I am rechecking that the database schema for 2.x and 3.x is the same
again. There was some recent drift between the two which would prevent
a 3.x installation from running against a database made with
FileSender 2.x. Since the drift is not large I am reinstating the
ability to migrate the database in that way.