Ability to store IdP metadata in a new table and link to transfer (#2234)
* Ability to store IdP metadata in a new table and link to transfer

This should help to dice up data for stats and work out which IdP the
transfers and users are coming from.

* hit idp remote md

* .

* allow metadata that doesnt have translations

* this is a bit more defensive

* cleanup

* Make the column names follow the standard for other columns

The auth_sp_idp_metadata_to_capture now allows for the SAML metadata
key to be something slightly different to the key used to store the
data in the local FileSender table column.

* avoid doing this too often

It is just a waste of electricity to expect the IdP metadata to change
for each transfer. Maybe 4 times a day is not so bad.

* Update Auths, update views, new cron script

Still getting there but things are moving in the right direction.
I should drop the old varchar from auths to clean that up too.

* Select the IDP entityID directly from new idps table.

* info line for md update. docs++

* c year

* allow extra storage

* ordering is important

* Update statistics_page.php

show nice name in the drop down but still use entityid as the value

* Update statistics_page.php

order by name

* New import-idp-metadata.php script

It is not as fast as it could be. But if you have only a few hundred
idp-remote entities and a relatively fast machine it should be plenty fast.

* Docs++, Change default to cron update only

There is a subtle difference in the idp scripts so that should be
mentioned in the README to help folks

It makes more sense to update the idps metadata using cron rather than
randomly in a transfer creation.

---------

Co-authored-by: Michael D'Silva <michael.dsilva@aarnet.edu.au>